Refactor engine_algebra.ex and in_memory_engine.ex
This commit refactors the code in engine_algebra.ex and in_memory_engine.ex. It introduces a new module, EngineFunctor.DeployDefinition, which represents the operation to deploy a new process definition. The code in engine_algebra.ex is updated to use this new module. Additionally, the in_memory_engine.ex module is modified to include a new function, do_run/2, which evaluates the DeployDefinition functor. This refactoring improves the organization and readability of the code.